requirements.txt
alabaster==0.7.11
asn1crypto==0.24.0
astroid==2.0.4
atomicwrites==1.1.5
attrs==18.1.0
Babel==2.6.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
coverage==4.5.1
cryptography==2.3.1
docutils==0.14
flake8==3.5.0
gprof2dot==2017.9.19
graphviz==0.9
hypothesis==3.66.2
idna==2.7
imagesize==0.7.0
isort==4.3.4
Jinja2==2.10
lazy-object-proxy==1.2.1
lxml==4.2.4
MarkupSafe==1.0
mccabe==0.6.1
more-itertools==4.3.0
mypy==0.620
packaging==17.1
pluggy==0.7.1
psutil==5.4.7
py==1.5.4
py-cpuinfo==4.0.0
pyasn1==0.4.4
pycodestyle==2.4.0
pycparser==2.18
pyflakes==2.0.0
Pygments==2.2.0
pylint==2.1.1
pyOpenSSL==18.0.0
pyparsing==2.2.0
PySocks==1.6.6
PyStemmer==1.3.0
pytest==3.7.4
pytest-benchmark==3.1.1
pytest-cov==2.5.1
pytest-profiling==1.6.0
pytz==2018.5
requests==2.19.1
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.7.8
sphinxcontrib-websupport==1.1.0
SQLAlchemy==1.2.13
typed-ast==1.1.0
urllib3==1.23
Whoosh==2.7.4
wrapt==1.10.11
